instantiateAccessGroups in add-agent-buttons.test.ts
