+++ test/unit/acl/add-agent-buttons.test.ts
@@ -1,17 +1,14 @@
 import { JSDOM } from 'jsdom'
 import { AddAgentButtons } from '../../../src/acl/add-agent-buttons'
-import { AccessGroups } from '../../../src/acl/access-groups'
+import { instantiateAccessGroups } from '../helpers/instantiateAccessGroups'
 
 jest.mock('rdflib')
 jest.mock('solid-auth-client')
 const dom = new JSDOM('<!DOCTYPE html><p>Hello world</p>').window.document
 
 function instantiateAddAgentButtons () {
-  return new AddAgentButtons({
-    controller: {
-      dom
-    }
-  } as AccessGroups)
+  const groupList = instantiateAccessGroups()
+  return new AddAgentButtons(groupList)
 }
 
 describe('AddAgentButtons', () => {
